# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/admin/users/{userId}:
    get:
      operationId: AdminController_getUserDetails
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetailsDTO"
        "403":
          description: Forbidden
        "404":
          description: User not found
      security: *a2
      summary: Retrieve detailed information about a specific user
      tags: *a3
    delete:
      operationId: AdminController_deleteUser
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSoftDeleteUserRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeleteUserResponseDTO"
        "403":
          description: Forbidden
        "404":
          description: User not found
      security: *a2
      summary: Legacy delete endpoint that forwards to soft delete
      tags: *a3
...
info:
  title: Serchat API
  description: The Serchat API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AdminUserDetailsDTO:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        login:
          type: string
        displayName:
          type: object
          nullable: true
        profilePicture:
          type: object
          nullable: true
        permissions:
          type: object
        createdAt:
          format: date-time
          type: string
        banExpiry:
          format: date-time
          type: string
        warningCount:
          type: number
        badges:
          type: array
          items:
            type: string
        bio:
          type: string
        pronouns:
          type: string
        banner:
          type: object
          nullable: true
        deletedAt:
          format: date-time
          type: string
        deletedReason:
          type: string
      required:
        - _id
        - username
        - login
        - displayName
        - profilePicture
        - permissions
        - createdAt
        - warningCount
        - badges
        - bio
        - pronouns
        - banner
...
    AdminSoftDeleteUserRequestDTO:
      type: object
      properties:
        reason:
          type: string
...
    AdminDeleteUserResponseDTO:
      type: object
      properties:
        message:
          type: string
        anonymizedUsername:
          type: string
      required:
        - message
        - anonymizedUsername
...